pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xiaodu-voice-control"
version = "0.1.0"
description = "Self-hosted XiaoDu voice control gateway for Home Assistant"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.32.0",
  "httpx>=0.28.0",
  "cryptography>=45.0.0",
  "pydantic>=2.9.0",
  "pydantic-settings>=2.6.0",
  "pyyaml>=6.0.2",
  "python-multipart>=0.0.12",
]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
xiaodu_voice_control = ["templates/*.html"]
